requirements.txt
alembic==1.12.1
alembic-postgresql-enum==1.0.2
annotated-types==0.6.0
anticaptchaofficial==1.0.58
anyio==3.7.1
arq==0.25.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.5
dadata==21.10.1
et-xmlfile==1.1.0
fastapi==0.104.1
h11==0.14.0
hiredis==2.3.2
httpcore==1.0.1
httpx==0.25.1
idna==3.4
lxml==4.9.3
Mako==1.3.0
MarkupSafe==2.1.3
numpy==1.26.1
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.3
Pillow==10.1.0
py==1.11.0
pycparser==2.21
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pypdfium2==4.23.1
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
python-telegram-bot==20.6
redis==5.0.1
requests==2.31.0
selenium==4.15.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.23
starlette==0.27.0
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
urllib3==2.0.7
uvicorn==0.24.0.post1
webdriver-manager==4.0.1
wsproto==1.2.0
